feat: Set up core project directory structure and files
